LayoutTests/imported/w3c/web-platform-tests/editing/other/insertparagraph-in-editing-host-cannot-have-div.tentative_host=span&white-space=pre-line&display=inline&command=insertText-expected.txt

a
b

PASS <span contenteditable style="display:inline; white-space:pre-line">a[]b</span>
PASS <span contenteditable style="display:inline; white-space:pre-line"><span style="white-space:normal">a[]b</span></span>
PASS <span contenteditable style="display:inline; white-space:pre-line"><span style="white-space:normal"><span style="white-space:pre-line">a[]b</span></span></span>
FAIL <span contenteditable style="display:inline; white-space:pre-line"><span style="display:block;white-space:normal">a[]b</span></span> assert_equals: expected "<span style=\"display:block\">a</span><span style=\"display:block\">b</span>" but got "<span style=\"display:block\">a\nb</span>"
FAIL <span contenteditable style="display:inline; white-space:pre-line"><div>a[]b</div></span> assert_equals: expected "<div>a</div><div>b</div>" but got "<div>a\nb</div>"
PASS <span contenteditable style="display:inline; white-space:pre-line"><div style="display:inline">a[]b</div></span>
PASS <span contenteditable style="display:inline; white-space:pre-line"><div style="display:inline-block">a[]b</div></span>
PASS <span contenteditable style="display:inline; white-space:pre-line"><div style="display:inline;white-space:normal">a[]b</div></span>
PASS <span contenteditable style="display:inline; white-space:pre-line"><div style="display:inline-block;white-space:normal">a[]b</div></span>
